backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=3.13
python-dotenv>=1.0
